OviSNAKE initial commit
The OviSNAKE is designed to do the automatic test by using the “Ovi Map
Middleware” and focus on the map images comparison. It can extract the
required data from a “txt” file then input these data to the middleware
and control the middleware to locate the specific place on the earth .
After that, the script opens a tool which named is “Sikuli” and compares
the expected map image with the actual image on the middleware . At
last, the script exports the compared results to a certain “txt” file.